requirements.txt
pandas>=2.1.1,<3.0
requests>=2.31.0,<3.0
python-dotenv>=1.0.0,<2.0
